requirements.txt
flask==0.10.1
python-dotenv==0.10.3
numpy==1.14.5
langdetect==1.0.7
talon==1.4.4
nltk==3.3
scikit_learn==0.19.2
theano==1.0.4
git+https://github.com/oppielabs/skipthoughts.git